2025-Nongae-Server/requirements.txt
flask==2.2.3
flask-socketio==5.3.3
requests==2.28.2
Pillow==9.4.0
python-engineio==4.4.1
python-socketio==5.8.0
eventlet==0.33.3 